FastMath: i_normalise, use long to prevent overflow.
This prevents i_normalise failing when x,y are large.

+++ src/Math/FastMath.cpp
@@ -237,7 +237,7 @@ void i_normalise(int &x,
   const unsigned m_max = std::max(abs(x), abs(y));
   if (!m_max)
     return;
-  const int mag = isqrt4(x*x+y*y);
+  const int mag = isqrt4((long)x*(long)x+(long)y*(long)y);
   x= (x<<NORMALISE_BITS)/mag;
   y= (y<<NORMALISE_BITS)/mag;
 }
